.......... [DEV-1161] added startup check to give error if an agent is already listening control socket, removed newlin from response

+++ go/src/zabbix/internal/agent/remotecontrol/remote.go
@@ -21,6 +21,7 @@ package remotecontrol
 
 import (
 	"bufio"
+	"fmt"
 	"io/ioutil"
 	"net"
 	"os"
@@ -49,7 +50,7 @@ func (c *Client) Request() (cmmand string) {
 }
 
 func (c *Client) Reply(response string) (err error) {
-	_, err = c.conn.Write([]byte(response + "\n"))
+	_, err = c.conn.Write([]byte(response))
 	c.conn.Close()
 	return
 }
@@ -88,6 +89,9 @@ func New(path string) (conn *Conn, err error) {
 	c := Conn{}
 	if path != "" {
 		if _, tmperr := os.Stat(path); !os.IsNotExist(tmperr) {
+			if _, err = SendCommand(path, "version"); err == nil {
+				return nil, fmt.Errorf("An agent is already using control socket %s", path)
+			}
 			if err = os.Remove(path); err != nil {
 				return
 			}
